Prefer PIC library when building with -c opt.
bazel build -c opt fails because -c opt makes cc_library rules generate
two outputs: a library built with -fPIC and a library built without
-fPIC. We only want the one with -fPIC but there is no way to force that
in a build file. (It can be forced with a command line option.)

This diff adds a pic_library rule to choose a .pic.a file if one is
available.